--- src/logic/README.md
@@ -1,1 +1,0 @@
-Here goes all the unit-testable functions.
